requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.94.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
python-multipart==0.0.26
sniffio==1.3.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
